+++ src/main/java/ru/averkiev/socialmediaapi/security/RefreshToken.java
@@ -25,7 +25,7 @@ public class RefreshToken {
      * @param createdAt дата создания токена.
      * @param expiresAt дата окончания действия токена.
      */
-    public RefreshToken(int userId, String refreshToken, LocalDate createdAt, LocalDate expiresAt) {
+    public RefreshToken(Long userId, String refreshToken, LocalDate createdAt, LocalDate expiresAt) {
         this.userId = userId;
         this.refreshToken = refreshToken;
         this.createdAt = createdAt;
@@ -36,11 +36,11 @@ public class RefreshToken {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private int id;
+    private Long id;
 
     /** Идентификатор пользователя владеющего токеном. */
     @Column(name = "user_id")
-    private int userId;
+    private Long userId;
 
     /** Строковое представления токена. */
     @Column(name = "token")
